+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Shared Pipes Component
 * Added `param` to `Trigger` definition.
 * Added safe pointer dereference function.
 * Added JSONTime type to handle time.Time in JSON output.
+* Added `GoToHCLString` function that converts a Go data structure to an HCL string.
 
 ## v1.4.3 [2024-07-12]